molden reader added, tests are next
